Add lblDownloaded to show download progress in updates
A new label `lblDownloaded` was added to `frmUpdateAvailable` to display download progress during updates. It is dynamically updated with the amount of data downloaded and total file size in KB. The label is initially hidden and toggled visible during the download process. The download destination was also changed to a temporary directory for safer file handling.